+++ package.json
@@ -89,6 +89,7 @@
     "getenv": "^1.0.0",
     "i18n-js": "^3.5.1",
     "jsog": "^1.0.7",
+    "lottie-react-native": "^3.5.0",
     "modal-enhanced-react-native-web": "^0.2.0",
     "moment": "^2.24.0",
     "node-gyp": "^6.1.0",
