+++ src/components/organisms/index.ts
@@ -1,4 +1,3 @@
-export { default as ModalAlertPublish } from './ModalAlertPublish';
 export { default as ModalLackPoint } from './ModalLackPoint';
 export { default as ModalAlertCorrection } from './ModalAlertCorrection';
 export { default as ModalBlock } from './ModalBlock';
